requirements.txt
albumentations==1.3.1
gradio==3.44.4
imageio==2.31.3
numpy==1.25.2
opencv_python==4.8.0.76
opencv_python_headless==4.8.0.76
Pillow==10.0.1
pycocotools==2.0.7
python_box==7.1.1
scikit_learn==1.3.0
scipy==1.11.2
torch==2.0.1+cpu
torchvision==0.15.2+cpu
transformers==4.33.0
